# Switch atomic.Value to atomic.Pointer for spanProcessorStates