This is a first look at switching `atomic.Value` to `atomic.Pointer`, so we have better type safety, and supposedly better readability.
I'm saying "supposedly" for better readability because for this specific case, it requires doing a lot of going from pointer to value, as a slice type can't be used from its pointer.

See #3918.