@pellared 2023-03-23
Is there any performance difference?